climate_monitor_pip_requirements.txt
influxdb==5.3.0
requests==2.21.0
urllib3==1.24.1
w1thermsensor==1.3.0
RPi.GPIO==0.7.0